#47 :sparkles: feat: 컴포넌트 수정
