+++ src/pages/donePage/ReviewBox/ReviewBox.tsx
@@ -1,3 +1,6 @@
+import { Children } from 'react';
+import { useParams } from 'react-router-dom';
+
 import { alignCenter, blueHighlight } from '@/pages/donePage/DoneHeader/DoneHeader.style';
 import PlaceList from '@/pages/donePage/Place/Place';
 import MoreReview from '@/pages/donePage/ReviewBox/MoreReviewBox';
@@ -19,18 +22,32 @@ import {
   mediumTextColor,
 } from '@/pages/donePage/ReviewBox/ReviewBox.style';
 import ReviewContents from '@/pages/donePage/ReviewContents/ReviewContents';
+import { TextColor } from '@/pages/donePage/ReviewContents/ReviewContents.style';
+import { ReviewDetail } from '@/pages/donePage/ReviewContents/ReviewContents.style';
+import { ImageStyle } from '@/pages/donePage/ReviewContents/ReviewContents.style';
+import { AlignCenter } from '@/pages/donePage/ReviewContents/ReviewContents.style';
+import { ReviewText } from '@/pages/donePage/ReviewContents/ReviewContents.style';
+import { ReviewImg } from '@/pages/donePage/ReviewContents/ReviewContents.style';
 
 import Button from '@/components/@common/Button/Button';
 import ContentBox from '@/components/@common/ContentBox/ContentBox';
 import Text from '@/components/@common/Text/Text';
 import Title from '@/components/@common/Title/Title';
 
+import { useReviewsInfoQuery } from '@/hooks/reservation/useReviewsInfo';
+
 import DoneBtn from '@/assets/svgs/done/done_btn_x.svg?react';
 import StarImg from '@/assets/svgs/done/done_ic_star_red.svg?react';
+import DoneImg from '@/assets/svgs/done/done_img.svg?react';
 import MemoIcon from '@/assets/svgs/done/memo.svg?react';
 import NextBtn from '@/assets/svgs/done/review_ic_next.svg?react';
 
 const ReviewList = () => {
+  const { reservationId } = useParams();
+
+  const ReviewInfo = useReviewsInfoQuery(+(reservationId || 0));
+  console.log('데이터 : ', ReviewInfo);
+
   return (
     <>
       <ContentBox
@@ -54,12 +71,31 @@ const ReviewList = () => {
             4.35
           </Title>
           <Text size={'large'} css={mediumTextColor}>
-            995개 평점 (769명)
+            2개 평점 (2명)
           </Text>
         </p>
         <div css={SlideWrapper}>
-          <ReviewContents />
-          <ReviewContents />
+          {ReviewInfo?.data.userReviewList.map((item) => (
+            <ReviewContents key={item.reviewId}>
+              <section css={ReviewDetail}>
+                <DoneImg css={ImageStyle} width={40} height={40}></DoneImg>
+                <div css={AlignCenter}>
+                  <Text size={'medium'}>{item?.name}</Text>
+                  <Text size={'small'} css={TextColor}>
+                    {item?.reviewDate}
+                  </Text>
+                </div>
+              </section>
+              <section css={SlideWrapper}>
+                <div css={AlignRow}>
+                  <Text size={'medium'} css={ReviewText}>
+                    {item?.content}
+                  </Text>
+                  <DoneImg width={48} height={48} css={ReviewImg}></DoneImg>
+                </div>
+              </section>
+            </ReviewContents>
+          ))}
         </div>
       </ContentBox>
       <ContentBox
